F1179: Fixed timing issue with animations
The problem would occur when someone would click on a fragment, and it
would load, but the animator wouldn't start before onPause().  Normally,
onPause() would finish the animator automatically, pushing the new
Fragment onto the stack.  But without an animator ready, it would delay
the animation until after rotation, at which point there were is a
timing issue with the menus.

The solution is to detect if an animation was run after a rotatation.
In those cases, we just end the animation as quickly as possible (as if
it were ended in onPause()).