F1310: Do not clear all callbacks on finish
There was a timing issue where callbacks were being cleared while new
requests for images were being made, such that PhoneSearchActivity.onDestroy()
was messing up loading images in LaunchFragment.  Now we check to make
sure that we're not finishing before we clear callbacks (as that is the only
safe time to be doing so).

We also now recycle all images on finishing PhoneSearchActivity.  Otherwise
we can end up with some thumbnails sticking around in memory for a long, long
time.