F1327: Combined hotel and fallback into same Adapter
The problem was occurring because we were replacing one adapter with
another, with unexpected consequences.  Now, we use a single adapter
that can be used as either a normal hotels results or fallback.  This
way it is treated the same regardless of what results we receive, and
thus should work for infinite scrolling.

There were other possible solutions (like trying to restore position)
but I decided this was easiest.