F1334: Moved reset of billing info sooner
What seemed to be happening was that onDestroy() would get called
after onCreate() of the next Activity - it would be delayed so much
(somehow) that it would screw around with the recreated Activity.

Now we do it in onPause(), guaranteeing it will be cleared nice and
early before you can reopen the Activity.