Default widget click now launches app
Google was complaining that clicking on the widget seemed to do nothing if
the hotel radar branding screen was up (and not a hotel).  Now it should launch
the app when clicked.  It will still use the old behavior when you click on
a hotel (i.e., launch to the details page).

This is a very simplistic fix for a fast release - it could be improved (like
sending the user to the search page) but we didn't want to deal with any flow
issues like we normally get from the widget.