Prepare repo for build change
We are going to modify the packageName for HA Jenkins to allow simultaneous install
of production and HA builds. This commit modifies the activity shortcut reference
to use the full packageName (as it won't be able to understand and find activities
with a modified packageName and the . based shortcut).

Additionally, modify the custom XML namespace for our custom attributes to use
the ADT 17 + approved "http://schemas.android.com/apk/res-auto" value. This barfs
out when modifiying the packageName in AndroidManifest.