Improve experience for physical keyboard users
Set focusable=true attribute on a bunch of ActionBar views such that the user
can navigate and select these views to perform our important ActionBar actions.

Many more topics will have to be addressed in order to fully support trackpad
usage, but this is some relatively minor changes in order to improve the flow
for now.