Fix HotelListFragment and HotelMapFragment timing issues
After introducing the ViewPager in a previous commit the order of
initialization changed drastically exposing some timing issues. Not only
that but since we use a FragmentPagerAdapter we relinquish control of
the fragment lifecycle. So I had to add a method that would notify that
the fragment was being attached when the FragmentPagerAdapter would
reuse fragments instead of creating new ones.