Add SupportMapFragment with newInstance call
Initial import of Dan's SupportMapFragment augmented with a queue of
camera events. I didn't know if it was necessary but it seemed possibly
useful.

This also provides SupportMapFragment.newInstance(). Previously if it
were to be called the wrong concrete type would have been instantiated
because it just calls up to the google SupportMapFragment
implementation.