User must go to itin page before we load itins.
This relates to commit 85c3df65a9775527d0fcdabc9ea4257f2f9340cb and should also only be temporary.
We dont want people working on other parts of the app to have to wait for the itins payload to download, so the user must first go to the itins page before we start loading itins (even if we are already logged in).
Seems like common sense but I over looked it in my earlier commit