87 - Speed up flight checkout animation.
It looks like a major culprit is the attach/detach Checkout being called from setCheckoutVisibility.
We removed the alpha animation on the CheckoutFragment container.
We removed the alpha animation on the top bar.
We altered the animation to slide up from below the fold of the screen, which makes up for our lack of alpha animations