Fixed timing issue on Spinner sections
The problem is that it is impossible to set a Spinner listener and
also manually set the Spinner's value without the listener firing the
listener (regardless of which order you set).  The only workaround
I've found in the past is to track when you're manually setting
the Spinner vs. letting the user set it.