285 & 207 Fix FB connect crashes.
I was not able to reproduce the crashes on my phones, but it looks likely they were caused by
listeners not being removed even when the fragment/activity were destroyed.
We also updated spacing so that all of the login form elements are visible on an N4.
We also removed some hardcoding, in favor of dimen values.