Fixed showAll() with zero visible markers
The showAll() method tries to animate the camera even if no LatLngs have
been added to the bounds.  This would cause a crash if you filtered the
results such that nothing is showing.

Now it will either not move at all (if there are no markers to show) or it will
show all the markers where they *would* have been (if they were visible)