Revert "Address audit concern, dynamic log tag"
This reverts commit 76e163335d2b23e6c0ad42d467ce27173a714677.

There was a concern addressed that we shouldn't do this. Essentially, log tags
are for devs to use and their utility comes outside of the code. If somebody has
a log tag filter set up, than refactoring of the class name will render this
filter moot. it is not worth making the log tag dynamic to introduce this problem