271 Fixed a crashing bug caused by timing issue.
The hotels overview was depending on Db.getBillingInfo() however, loading that same billing info was happening in a background thread, so sometimes the loading wouldnt be finished when the ui called Db.getBillingInfo and then we would crash.

We moved the loading of checkout data to seperate object so we could share code between flights and hotels.
We are now thread safe and happy.