Properly align BlurredBackgroundFragment bg in overlay mode
The bg was not accounting for the top bar, which would be great if we
weren't always showing it.  By taking the top bar into account, we
benefit properly aligning with the other activities (without them
having to use special overlay modes)

In the process I had to fix some math in FadingImageView.  rect.top
was always 0 before, so I had it the wrong way around in testing.  I
should have been subtracting instead of adding.