Parse out required payment fields for hotels.
Currently, the architecture for this reflects the ExpediaPointOfSaleConfig.json
file and the values it returns. It is uncertain how this behavior will change in
the future and how we will represent in the file, so I have gone the simplest
route as possible for now.

If we start having more complex usages of required payment fields, the backing
data structures may want to be rethought.

Auditors: dlew