ITIN - Flight Details - Early version of flight leg info.
This version basically just gets the views on screen and draws the flight lines.
However it is broken due to the fact that we are currently showing multiple flight legs on one card, even if they are different parts of a round trip. This will be fixed after the adapter gets fixed/updated