ITIN - Flight Details - Display flight insurance with link.
We now parse and display flight insurance. This is pretty sketchy because the insurance is tied to the Trip and the Trip
could have multiple flights. Which one is insured??? We will maybe never know, but for now if the Trip has insurance,
and it is of lineOfBusiness=AIR then we display the insurance with the flight(s) in the trip.