356 Hotels Calendar/SearchParams weirdness
During the internal date refactor, some buggy code was introduced pertaining to
the setting of the SearchParams checkin/checkout dates. java.util.Calendar uses
zero index month, and was causing the date comparison to go awry.

Auditors: pardom, dlew

Other questions:

1. Why does setCheckInDate() validate the checkOut date (and vice versa)?
2. What is the test plan for 121 (internal date refactor)? I know QA has asked me about how to
test the story. The story is still in ready for test but this defect is most certainly
related to #121.