374: Normalized all dates in hotels
This fixes a MAJOR bug wherein the dates in the hotels searches were off
by one day in some timezones!  This problem is a result of us switching to
Date internally.  Now that we're no longer storing the timezone, we need
to work with the user's current timezone instead of normalizing to UTC
for display (since the Calendar returned by Date.getCalendar() is not in UTC)

Another possible solution would be to have Date.getCalendar() return in UTC,
but overall it will be simpler if we remove TimeZones from the app entirely,
so I chose to go with this route.