381 Create map marker if the property is new
This was happening because we initially set the properties we know about
from SearchResponse.getFilteredAndSortedProperties().

The problem with this is when we get filtered and start the
PhoneSearchActivity again we only know about the filtered properties,
not the entire set. So if we see more properties than we know about we
add the new ones to our set of known properties.