394 NPE crash
A BackgroundDownloader callback from a previous activity was being invoked and
crashing the app because the widgets the callback interacted with had been
nulled out. It seems like the unregistering of the callback was being invoked AFTER
the next activity had been launched. I moved around the timing of the cleanUp method
to ensure that it is called BEFORE the cleanUpOnStop is invoked, preventing the crash.

This is my theory anyway...anybody care to comment?