ITIN - GO AWAY YOU STUPID WHITE LINE!
It should be pointed out that before I discovered this one line fix, I first spent > 1 hour writting new data set observers
to listen for when the dataset reached 0 and then set the visibility of the background line to gone. This worked on 4.x
but not 2.x and I was trying everything, setting the line to gone, invalidating all of the views, etc. Nothing worked.
My computer then crashed, and somehow boned my configs so I spent another 45 minutes to get the stupid thing to build. Awesome.

Then I just set the background color of the empty view and it solved my problem. I hope my anguish is amusing to someone.