ITIN - Moved itinerary loading to the login fragment.
This commit is a bit strange. The LoginFragment is already too big (read too complicated) and I imagined it being forseeable that we may want to do other things on login in the future.
Thus the LoginExtender interface was born. This lets us do work after the login finishes, and have a place to tie into the UI if need be.
In the future I would like for the LoginFragment to have a collection of LoginExtenders that run in sequence after login.

ItineraryManager Change: I updated the manager to copy the listener list when iterating over it. This allows us to do things like remove a listener fromt he manager from a listener callback.