324 Itin Omniture tracking
Add the tracking calls according to the newly updated spec. Things yet to complete:

- Invocation of reload tracking calls (reload button does not exist in code yet)
- Better place the share tracking calls to include long vs. short. This would put
the placement of the tracking calls in SocialMessageChooserDialogFragment in
order to determine long vs. short. Unfortunately, the creation of this dialog
fragment occurs in ItinItemListFragment, which does not appear to have a convenient
way to determine which itin card is which (and invoke the correct tracking  call).
This might require changing of the the OnItinCardClickListener interface to include
card type, or, maybe there is another way that I am not aware of.

What do you guys think, Pardo/Joel?

Auditors: pardom, jdrotos