ITIN - dehose 2.x #2 - Animator me this, animator me that.
We removed the use of layout animations, as they are the old style of animations, and it probably wasnt helping much of anything to mix and match.
We now return animatorSets from ItinCards so that we can kick off the animations all from one place (and one time)
Added a bunch of logging that will be removed after this thing is where we want it.