ITIN - dehose 2.x #3 - Our animation may be ugly on older phones, but we can atleast display data now.
Note we added a new set of row types to the adapter - the Details state. This solves a bug where we would be expanding all of our Flight rows at once, and boy was it ugly.
We also solved a scrolling issue, by scrolling during the animation only on newer devices.