ITIN - 550 - Expedia itin number and confirmation codes added to all card types where available. Also giant refactor.
The giant refactor moves a bunch of duplicate code form the individual ItinCard sub types into ItinCard.java.
This should make maintenance quite a bit easier, and allow for a more consistant experience through the card types.
The code for adding the common gui elements takes a container as a param, which things get added to.
This way the ItinCard subtypes can call it if they want, or do a custom solution if so required by design or whatever else.
Functionality shouldnt really be changing (except for the addition of itinerary # to details views).