An autoscroll to the correct row update.
We changed the behavior slightly. Now we only autoscroll if the list used to be empty. So if we have looked at the list, coming back to it shouldn't make us autoscroll.
This is expected behavior, I dont want the thing jumping around unless if it is loading up for the first time, at which time I will want the most relevant item.