Created ItinContentGenerators
These allow you to generate the itinerary card content without having it
be tied to the framework that actually puts them onto the screen.

As a result, there is now only one ItinCard which uses an
ItinContentGenerator (based on the ItinCardData type) to generate all the
content inside of an ItinCard (instead of one ItinCard per type).