230 add reload button to expanded card details
I am not fully satisfied with my implementation here, but hey, it seems to work.

- It feels a little bit wrong to be broadcasting messages from ItinCardDataAdapter to be received
by ItinListView. This avoids setting up an ItinListView reference in ItinCardData Adapter, though.
- Sending a message on every onTripUpdated and onTripUpdateFailed feels even more wrong, but it gets
around the problem of ItinCardDataAdapter not being knowledgeable of the card that is expanded (and
ultimately, notifies the ItinListView on every update such that the ListView can determine if a card
 needs to be redrawn).

Anyway, Joel/Dan/etc.., I want to hear your feedback.