549 - NPE protect on WIFI disconnect
This solution is in similar vein to a commit of Joel's from earlier... protect at all costs against
null content in the ItinCardData set of classes. I was thinking of creating an interface for the
ItinCardData set of classes, "ContainsContent" that each subclass can implement and thus we could nip
this problem off at the bud earlier (perhaps inside of ItinCardDataAdapter.syncWithManager by only adding
cards with content to the adapter). Ultimately, this should be handled higher up in the chain
though (maybe in the parser), so I have just done this quick fix for now to make the app less crashy.