ITIN - 600 - Fix expanding cards at bottom of list in 2.x
This was waaaay harder to figure out than it should have been.
So our problem was that if we clicked the last rows in the listview, the rows would not have room to expand because
we cant scroll to things that dont have space below them. Adding a footer view seemed like a good option.
However it takes the footerview another layout pass before it is drawn (and measured) and we want our animation to run now.
ViewTreeObservers onGlobalLayout did not help in this case because though the view was layed out, it seemed that
it wasnt for sure drawn.

We ended up with this sort of basterdized recursive call to showDetails() that adds a footerview and keeps trying to
scroll to the point that puts our selected row at the top of the scrollview (which cant happen unless our footerview is in the layout).

For future reference:
1) To add a footer view you need to do so atleast once before the adapter is set (subsequent footer views can be added whenever)
2) Adding a footer view will cause a re-measure of our list rows, so we must wait until it is in the layout before starting our animations.
3) I spent a long time trying to determine if the footerview was needed or not (e.g. how much space do the rows below the selected row occupy)
   this was where most of my time went. It turns out that there is not a great way to do this. Instead trying to scroll to the view we want
   was way more effective because it will succeed right away if we have enough views below the selected view, otherwise it wont succeed until
   our footerview is happily in the layout.