655 Added missing street address and zipcode for Car Rental share via text.
This will be a temporary fix.
As per my conversation with Scott, we will have to add a smarter way of showing address depending on the time the message is composed.
i.e if the user shares BEFORE pickup, we should use the pickup address
if they share AFTER pickup, we should use the dropoff address