Properly add guest itinerary after guest booking
1. On Hotels path, make sure to reference the itinNumber from BookingResponse (not CreateTripResponse
which is only invoked when applying a coupon).
2. On Flights path, make sure to send the itinNumber as opposed to itinId.
3. Make sure to init mTrips of ItinManager when adding guest trip. I've seen this be null, so make sure to
create data structure before attempting to add to it.

Coincidentally, this should clean up ancient defect #464 which appears to be referencing some other older
crashers, but this commit effectively makes it so guest bookings are automatically added to the itin screen.

UNFORTUNATELY, I can't get TRIP_DETAILS call to go through properly on trunk, despite running it against some
live bookings. So, I am not pushing the defect along until I can actually test it out and make sure it works.