668: Split phone into two fields
We're supposed to show both phone #s if we get them both, but the code was
only showing one or the other.

Also, we now use the local phone # if given a choice instead of toll free
(as that is what iOS is doing).

In addition we only show the phone # if it's available.  If the data is not
available then we do not show anything.