629 protect against NPE on logout (band-aid)
Look at the log for this defect and you'll see that LaunchActivity is touching
a public method of ItinListFragment wherein the fragment invokes getActivity()
only to receive null... stop this from happening by validating the fragment null
check to also check for Fragment.isAdded().

Ultimately, I'm not entirely sure why the Fragment was not added at the Activity's
menu override's time. My hunch is that this defect was voided when Joel fixed the
Fragment references wrt. ViewPager. Or maybe it is a weird timing issue related to
the menu overrides being invoked before Fragment can become added. Either way,
protect against the NPE.

Also, added a utility method to Ui.java to standardize our null/isAdded check we
seem to always be performing (good idea, Brennan). Ui.java might not be the best
place for this method as it is not using an generics (my generics foo is non-existant,
but I don't think we require it in this case unless we want to make the fragment
parameter more lenient which shouldn't be needed in this case).