675: Logging out of itins deselects on tablet
This was a bit of an involved change; instead of the Fragments having
dialog listeners I moved the listeners up to the Activities (which then
inform the fragments as necessary).  This allows an Activity to inform
multiple fragments of state change.

I also cleaned up the logout dialog (as it had some unused code in it
that was just confusing the issue).