Cookies - 686 - Dont store cookies of cancelled download.
This change scares me a bit, as should any change that messes with EB cookies.
The problem identified by defect686 seems to be caused by the user getting to the checkout page,
the refresh login call getting started, the user calling logout, but regardless of refresh signin
request being cancelled, we still call save cookies afterwards, and the cookies we save were put
into memory before the logout was called. We now only save cookies if a request was not canceled.