706 allow user to book hotel on tablet
The hotel checkout flow on tablet has never had to deal with the case where
you can login from somewhere other than its own booking screen. Make sure to
setup the booking form properly if already logged in (as you can be already
logged in from the itin screen).

I was going for the one line fix here. This fragment is generally a mess, but,
this seemed like the most logical fix.