730: Add more logic for summary view
Before, if you had an itinerary currently in progress and another itinerary
far in the future, it would show that one in the summary view instead
of the one currently in progress.  This is because our current algorithm
focuses on either itins that happen TODAY or itins that have YET TO START.

Now we take into account itins that are in progress (but did not start
today).  If they end before the next itin start and the next itin starts
more than three hours from now, we use that in the summary view instead.

This easily fixes the problem stated above.  It also errs on the side of
earlier summary views when the end date of one is on the start date of
another (e.g., you're checking out of your hotel and jumping on a flight).