Move ItinCardData generation into ItineraryManager
The problem is that a crucial piece of info (the list of ItinCardData)
is controlled by an adapter, which is inside of a listview, which is inside
of a fragment.  This means that if anyone else wants to use this data
they have to go through all those classes to get to it, and also that the
view is controlling the data.

This solution results in the same behavior as before, except that you can now
retrieve the same ItinCardData that the list is using elsewhere.