Better handle removed trips in the list
If there is an expanded details card and the trip happens to be removed
while the card is expanded (could happen if you perform a deep refresh while
the card is expanded, and the refresh of the trip results in its removal, say,
if you cancelled it on desktop while viewing expanded in app).

This also seems to be enough to move along #626 which Daniel already pushed a fix
for this earlier in the week.