734 don't crash on BookingOverviewActivity
There were two problems

1. We were saving a boolean in onSaveInstanceState that tracked whether or not
the BillingInfo had been loaded from Db. Unfortunately, this piece of data persists
after the app is killed in the background but the Db does not. Remove this boolean
and just check the Db instead.

2. Once this crash was fixed, another issue was that Db.selectedProperty gets flushed
out when app is killed in background. Make sure to exit the activity which ultimately
sends the user back to the search screen in this case.