770 allow more than 5 chars on postal code field
We were limiting to 5 on hotels checkout for tablet. Apparently there
exist postal codes in Canada that have more than 5 characters. I
was thinking of limiting the US region to 5 chars but I for all I know
there could be a postal code in the US that is more 5 chars.