736: Added mdpi bg asset
It's really hard to tell why the app is OOMing (the HPROF dumps at
 the time of the crash show it only using half of its max memory!),
but it points towards the BG ImageView as the source of the crash
each time.  As such, I suspect it was the conversion from xhdpi to
mdpi that was causing problems; so by including an mdpi version of the
bg, it does not seem to OOM anymore.