750: Fixed MapFragment broken-ness on rotation
Having two overlaid MapFragments does NOT work.  Had to remove that.
Tried a lot of things, but ultimately we were using undefined behavior
and that is just asking for trouble.

The new solution zooms in quickly, measures, then zooms back out again.
This causes a slight flicker to occur, which is noticeable.  However,
it is far more preferable than any of the alternatives at the moment.