835 enforce LocationFinder timeout on Launch screen
Certain devices that do not have a sim card have a har time finding a Location
and will in fact just sit there for ever, never finding a location, yet never
returning that it has failed. In this case, the location is not super crucial
because we have fallback logic. IT is important to show pretty pictures on our
launch screen.