748: Make it harder to overlap searches
The problem was that the search response was taking a while to process (aka,
do the initial filter).  This would mean that it was relatively easy to start
a search at the same time as the results for the last search were processing.

Now most of the processing is done on another thread which is much easier to
cancel.  This doesn't mean the problem is gone entirely, but it should now be
MUCH harder to run into this problem.