290 - Update some strings.
Note the originals of all the strings I updated are still being used elsewhere in the app.
Also adding a new string for Select Payment seems dumb, but I asked george and they want
sentence case, and the other use of that string is for a heading which makes sense to have
words capitalized, so here we are.