467 Facebook install tracking.
The recommendation is to put the call to publishInstallAsync in onResume of every activity.
We put it in SearchActivity under the premise that it will get called when the user starts the app,
the docs were not clear as to why they wanted it in onResume() but a quick look at the code doesnt reveal anything
that shouldnt be called in onCreate, so this should be ok. Well need to have a look at the FB
console to ensure this is functioning properly.