Initial full wallet request & booking
It currently defers to the HotelBookingActivity for doing the actual
booking (the reason being that I wanted to avoid duplicating a lot
of the booking handling code).  We could move it away if we want to
later (at the price of having to re-engineer the process, perhaps
into its own invisible Fragment).