922 - Fix flight tablet backstack issues.
Because we are doing a little magic to end up back at the flight search activity when coming back
from flight results, we also need to do a little magic to get back to our start activity from the
flight search activity. Basically finish() doesnt work on the tablet side because SearchFragmentActivity
implements ActivityKillReceiver (which is good because it hogs memeory), but bad because it is not there to go
back to, so we just intent our way back.