Created LoadWalletFragment
While implementing the Flights portion of Google Wallet I noticed that
a *ton* of code could literally be copy/pasted between the two
implementations.  As such, I decided it would be a lot better to create
a version of WalletFragment that handles most of it for you.

There are also some enhancements in here, like renaming methods and
unifying some logic in the parts that need to be written in each subclass
so that they don't all have to implement it.

Each of these subclasses to WalletFragment should have their own specific
purpose, as the flows are handled substantially differently depending
on what you're trying to do (load, change or get full wallet).