592 show fare name on flight details
DO NOT TEST YET. The fareName is only returned for low cost carriers which we do not yet support.
Low cost carrier support is supposedly coming soon on trunk. I have decided to push now in interest
of preventing the local branch from becoming stale (and the work is done, just not yet ready for test).