907 Fix itin path line problems.
So basically the line was just showing on logout, and that was wrong, but it also
wasnt showing properly in a lot of other cases. As in it went beyond where it was
supposed to stop, started above where it was supposed to start, and cached strangely.

This should be fix now as we have the listview itself draw the line.
As an added bonus, we should see a performance improvement, because we are no longer
needing to a) animate the alpha of the path view, b) layout the path view (which happened a lot)
and c) we no longer have to listen for scroll events at the fragment level.