showBaggageFeesNotIncluded -> hasBagFee
API team had a change of heart on what to name the field. I decided to go ahead
and refactor its representation in our objects. I am not certain the new name is
more clear but at the very least it is more succint. Might as well closely match
the terminology that the API uses.