Improve behavior when pre-authed on mult. accts
It was spinning indefinitely (because it would be pre-authed, but would
also require resolution).

In addition, it would spinner each time you resume the page (either
through rotation or turning the phone on/off).  Now it will only spinner
the first time; once it figures out you need a resolution, it will stop
trying to assume you should check preauth.