Omniture code organization refactor
Remove TrackingUtils, which was a class that contained methods of a few different flavors. It
contained public methods to be used for tracking simple events and also internal helper methods
for preparing the tracking object. Its presence along with OmnitureTracking was just confusing.

TrackingUtils is now gone and its methods are in OmnitureTracking. The overall goal is to make it
such that OmnitureTracking is the one-stop shop for all Omniture related things. It contains a host
of public methods for use inside of Fragments and Activities and also private helper methods. Perhaps
OmnitureTracking is becoming too bloated, but I feel that consolidating these two source files does
more good than harm.