452 Fix NPE backing out of flights.
It appears that there was a race condition where we would be creating
animators before onCreateView and things would not be ready yet. We now
just do some safety checks, but because the FlightSEarchResultsActivity.onFragmentLoaded
method will get called again after onCreateView it should fix itself after
the veiws are loaded. Even if not, the animations just get a bit ugly, but still function.