1070 If we hide a view, we should update nextFocus views.
We were hiding the middle name, but never updating views who have nextFocus*
set to the middle name. Now if we remove a view we make sure that we pass the
nextFocus views along. e.g. if firstname.nextFocusDown == middleName, and middleName.nextFocusDown == lastName
then when we hide middle name we set firstname.nextFocusDown = lastName.