Crank up number of proguard optimization passes
Because if it can remove instructions and what have you, let it keep
doing it. It is smart enough to detect when it is has not done any work,
however it is not smart enough to detect any fixpoint. But after 7
passes things seem to be good so I gave it a little margin.