Itin card deep refresh disappearing act bug
There was a bug with deep refresh where if the refresh resulted in its removal,
and the card was expanded, then the ListView would crash or behave very strangely.

This took WAY too long to track down and my brain sort of turned into mush as a result.
I think perhaps there could be some consolidation of the mDetailsPosition that is tracked
in both ItinListView and ItinCardDataAdapter, although, I think I am done trying to refactor
Itin related stuff for now.