867: Fixed padding of plane window view
It turns out that RelativeLayout's centerInParent truly centers in the parent,
disregarding padding.  This does not mesh well with our overlay mode, which
requires adding extra padding to the top/bottom of fragments underneath the
action bar.

I've added a wrapper FrameLayout that will allow us to keep using padding in
the same manner as before.