1074 fix honeycomb settings page crash
Ultimately, this was introduced with Doug's recent fix for 574:
0ce9fd53a9cb5be03c098ccf8dc42c2292af8ac1, where he updated a style's
parent for values-v11 to be standard Android holo theme. This is bad
news because the same style was being used for an Activity and ABS
threw an exception. I fixed it by declaring a whole new style to use
for the 574 fix, rather than using Theme.Dialog.Light. This seems
to allow the fixes for both 574 and 1074 to co-exist.

Will someone that has a sense of style tell me if my fix is reasonable?