1013 Itin list scroll bug, and onTouchEvent refactor.
So this change is pretty big, but it should result in better touch behavior in ItinList.
Basically we were passing a lot of touch events to ItinCard, but not compensating well when
touches moved outside of cards or to new cards. Now we should be dispatching touch events
to the correct views at the correct times. Note: I tried to REDUCE the amount of code here,
but ultimately controlling the onTouch stack is non-trivial, and if we are overriding, we
may as well take full responsibility and handle all cases.