1003 Omniture coupon tracking
I'm not really sure about about the value for evar24 on Coupon Applied event
because all of the coupons I've tried (test coupons on integration and expedia-escape
coupons issued on basecamp) have an empty promo description. This seems like the most
likely place for it though. To verify with Omniture team.