Fix navigation issue on tablet wrt flights and itins
Auditors: jdrotos

This clashes with your recent fix for #922. ActivityKillReceiver was never intended
to be implemented on the "base" activity. If it turns out the memory impact is substantial
enough to warrant the activity being destroyed, we can add in some more NavUtils goodness,
although, I opted for the simple route for now. Also, we could solve the SearchFragmentActivity
memory concerns in another way, rather than destroying the Activity.