Add actionbar spacer view.
We're using the overlay actionbar mode to fix the itin expand
animation. The viewpager used to be pushed down by a spacer view,
but that added a layer of complexity. This was fixed in 70439d047
but I forgot to add a spacer view to the launch fragment as well
as the itin list fragment.