1106: Don't continue onCreate() if reorienting
What was happening was that we were dynamically setting the requested
orientation to portrait, but then continuing to run the rest of onCreate().
That's fine if we're already in portrait, but if we were in another orientation
then the Activity needs to be recreated.  However, we'd keep running more
tasks that occurred in onCreate() even in that situation (which included some
tasks in other Threads, which would eventually NPE).

Now we break out of onCreate() early if we're going to be re-orienting the
Activity.