911 prevent domestic flight searches in India
Now, once we send the user to the PlaneWindowView (and the FlightSearch is getting prepared to
kick off) we check to see if the search is a domestic India flight. If so, we display a dialog
saying that it is not supported yet and finish() back to the FlightSearch screen, otherwise, we
just kick off the flight search as usual.

Auditors: dlew