1183: Vastly improved flight search edit text anim
The animation was barely one before.  Now we use a number of tricky things to
avoid having to request layout mid-animation, which smooths the animation
tremendously.

Unfortunately, the utilities being used for the animation are API 11+.  Since
there was practically no animation before, I've simply removed it for pre-11
devices.  This seems to smooth the app as well.