Save only FlightSearchState on state change
Before we were saving the *entire* FlightSearch each time it changed
state; this was very inefficient because the FlightSearch itself could
be upwards of a megabyte or more.  It not only caused memory issues
but also performance issues.

Now we save only the state during a state change, which is *far*
smaller payload and takes almost no effort.