1184: Removed PVH from NOA implementation
NineOldAndroids has a bug in it, such that it does not proxy
ObjectAnimator for 2.3- when using PropertyValuesHolder.  I do not
want to sacrifice the performance gained in later versions of
Android, so we now use both methods (depending on the version).

If it is one day fixed, it will be here:
https://github.com/JakeWharton/NineOldAndroids/issues/48