PushNotifications - update new GCM regId sync call.
Before we were forcing a full ItineraryManager sync when we got a new GCM registrationId,
now we just do the push notification stuff. Note that this will typically only ever get called
the first time the user views itins, so not a huge deal, but a nice little performance boost.