PushNotifications - dont send dup payloads.
Before we were sending push notification registration requests everytime we did an
ItineraryManager sync. Typically the data would stay the same, and so this would be
a pointless network call, and it would cause undue server load (millions of users, syncing
every few minutes). So now we cache our payload, and only send data to the server if our payload
has changed.