Split up FlightConfirmationFragment
The idea is to isolate the parts of the page that differ from the
parts that stay the same (between all the confirmation pages).  Now
the code is split up in such a way that the actions themselves can
be defined for each (as well as the header) but some of the basic
layout components can remain relatively stable.