PushNotification memory optimization.
We were storing our registration payloads in memory to prevent ourselves from
sending the same payload twice in a row. Unfortunately the last thing we needed
was more JSONObjects sitting around in memory, so now we hash our payload to a nice
small string and use that for our book keeping.