137 Hotels side.
We now show the tooltip popup on the hotels side of the app.
This change includes allowing the user to have no dates, or just
one date selected on the calendarDatePicker (without breaking everything).
We also added calendar header instruction text on the hotels side.
We also allow the user to select long date ranges, and display a popup
if the date range exceeds maximum range (28 days).