Refactor Flights/HotelsRulesFragment layouts to styles
The TextViews + divider Views inside are all very similar and warranted a styles
unification. Cuts down on code duplication and will make the layouts easier to change
in the future.

FUTURE TODO: Generate these layouts dynamically rather than defining a layout file at all.
There is a decent amount of code branches that determine which TextViews are present and in
turn should be shown on the screen. It could just be worth addings these views programmatically
via TextView constructor or inflating a simple TextView in Java code, who knows.